LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/paused_true_during_pause-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.
CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.
spec reference

  

Harness Error (FAIL), message = The operation was aborted.

PASS audio events - paused property 
PASS audio.paused should be true during pause event 
PASS video events - paused property 
PASS video.paused should be true during pause event 
